# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "知行 ZhiXing Personal Agent",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/filter/rules": {
      "get": {
        "summary": "Get Filter Rules",
        "operationId": "get_filter_rules_api_filter_rules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Save Filter Rules",
        "operationId": "save_filter_rules_api_filter_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRulesInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/filter/box": {
      "get": {
        "summary": "List Filter Box",
        "operationId": "list_filter_box_api_filter_box_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "filtered",
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/filter/box/{filter_id}/release": {
      "post": {
        "summary": "Release Filter Item",
        "operationId": "release_filter_item_api_filter_box__filter_id__release_post",
        "parameters": [
          {
            "name": "filter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/filter/classify": {
      "post": {
        "summary": "Classify Message Api",
        "operationId": "classify_message_api_api_filter_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/filter/labels": {
      "post": {
        "summary": "Add Labeled Sample",
        "operationId": "add_labeled_sample_api_filter_labels_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabeledSampleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Labeled Samples",
        "operationId": "list_labeled_samples_api_filter_labels_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/filter/shadow": {
      "post": {
        "summary": "Shadow Classify Api",
        "operationId": "shadow_classify_api_api_filter_shadow_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pricing/{profile_id}": {
      "post": {
        "summary": "Update Pricing",
        "operationId": "update_pricing_api_pricing__profile_id__post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClassifyInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 30000,
            "minLength": 1,
            "title": "Text"
          },
          "sender_id": {
            "type": "string",
            "title": "Sender Id",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ClassifyInput"
      },
...
      "FilterRulesInput": {
        "properties": {
          "whitelist_senders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Whitelist Senders"
          },
          "blacklist_senders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blacklist Senders"
          },
          "blacklist_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blacklist Domains"
          },
          "subject_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subject Keywords"
          },
          "content_keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Keywords"
          },
          "auto_filter_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Auto Filter Categories"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "title": "FilterRulesInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabeledSampleInput": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 30000,
            "minLength": 1,
            "title": "Text"
          },
          "sender_id": {
            "type": "string",
            "title": "Sender Id",
            "default": ""
          },
          "expected_category": {
            "type": "string",
            "title": "Expected Category"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "manual"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "text",
          "expected_category"
        ],
        "title": "LabeledSampleInput"
      },
...
      "PricingOverrides": {
        "properties": {
          "input": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input"
          },
          "cached_input": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Input"
          },
          "output": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output"
          },
          "peak_input": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Input"
          },
          "peak_cached_input": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Cached Input"
          },
          "peak_output": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,23}0*$)\\d{0,10}\\.\\d{0,12}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Output"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{3}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "peak_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Enabled"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "peak_windows": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 12
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Windows"
          },
          "peak_weekdays": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 7
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Weekdays"
          },
          "holiday_dates": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Holiday Dates"
          },
          "calendar_years": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "maxItems": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Years"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PricingOverrides"
      },
      "PricingProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "default_version": {
            "type": "string",
            "title": "Default Version"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "calendar_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendar Source"
          },
          "defaults": {
            "additionalProperties": true,
            "type": "object",
            "title": "Defaults"
          },
          "overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Overrides"
          },
          "effective": {
            "additionalProperties": true,
            "type": "object",
            "title": "Effective"
          },
          "field_sources": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Field Sources"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "active",
          "version",
          "default_version",
          "source",
          "calendar_source",
          "defaults",
          "overrides",
          "effective",
          "field_sources"
        ],
        "title": "PricingProfile"
      },
      "PricingUpdate": {
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Version"
          },
          "overrides": {
            "$ref": "#/components/schemas/PricingOverrides"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "overrides"
        ],
        "title": "PricingUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}